pyproject.toml
[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "q_calc"
version = "1.0.0"
repository = "about:blank"
license = "MIT"
description = "small calculator that supports quaternions"
authors = [ "Ian <shiyan3@msu.edu>" ]

  [tool.poetry.dependencies]
  python = "^3.9.0"
  black = "^23.3.0"
  isort = "^5.12.0"
  numpy-quaternion = "^2022.4.3"
  ply = "^3.11"
  prompt-toolkit = "^3.0.39"

  [tool.poetry.scripts]
  qalc = "q_calc.parser:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
